feat: Add BukkitLogFactory for plugin logging
This commit adds a new class, `BukkitLogFactory`, which extends `LogFactory` from the `cn.hutool.log` package. This class is responsible for creating instances of `BukkitLog` for logging in the plugin. It overrides the `createLog` methods to return instances of `BukkitLog` based on the provided name or class. This change improves the logging functionality in the plugin by providing a custom logger implementation specifically designed for Bukkit.